requirements.txt
streamlit>=1.28.0
pdfplumber>=0.9.0
pandas>=2.0.0
openpyxl>=3.1.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
numpy>=1.24.0
python-dateutil>=2.8.2 